// extensions/htlc/besu-erc20/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface InitializeRequest {
    /**
     * connectorId for the connector besu plugin
     * @type {string}
     * @memberof InitializeRequest
     */
    connectorId: string;
    /**
     * 
     * @type {Web3SigningCredential}
     * @memberof InitializeRequest
     */
    web3SigningCredential: Web3SigningCredential;
    /**
     * 
     * @type {number}
     * @memberof InitializeRequest
     */
    gas?: number;
}
// ...
export interface RefundRequest {
    /**
     * Contract htlc id for refund
     * @type {string}
     * @memberof RefundRequest
     */
    id: string;
    /**
     * Contract address
     * @type {string}
     * @memberof RefundRequest
     */
    contractAddress: string;
    /**
     * 
     * @type {Web3SigningCredential}
     * @memberof RefundRequest
     */
    web3SigningCredential: Web3SigningCredential;
    /**
     * connectorId for the connector besu plugin
     * @type {string}
     * @memberof RefundRequest
     */
    connectorId: string;
    /**
     * 
     * @type {string | number}
     * @memberof RefundRequest
     */
    gas?: string | number;
}
/**
 * 
 * @export
 * @interface RunTransactionResponse
 */
export interface RunTransactionResponse {
    /**
     * 
     * @type {Web3TransactionReceipt}
     * @memberof RunTransactionResponse
     */
    transactionReceipt: Web3TransactionReceipt;
}
/**
 * @type Web3SigningCredential
 * @export
 */
export type Web3SigningCredential = Web3SigningCredentialCactusKeychainRef | Web3SigningCredentialNone | Web3SigningCredentialPrivateKeyHex;

/**
 * 
 * @export
 * @interface Web3SigningCredentialCactusKeychainRef
 */
export interface Web3SigningCredentialCactusKeychainRef {
    /**
     * 
     * @type {Web3SigningCredentialType}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    type: Web3SigningCredentialType;
    /**
     * The ethereum account (public key) that the credential  belongs to. Basically the username in the traditional  terminology of authentication.
     * @type {string}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    ethAccount: string;
    /**
     * The key to use when looking up the the keychain entry holding the secret pointed to by the  keychainEntryKey parameter.
     * @type {string}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    keychainEntryKey: string;
    /**
     * The keychain ID to use when looking up the the keychain plugin instance that will be used to retrieve the secret pointed to by the keychainEntryKey parameter.
     * @type {string}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    keychainId: string;
}
/**
 * Using this denotes that there is no signing required because the transaction is pre-signed.
 * @export
 * @interface Web3SigningCredentialNone
 */
export interface Web3SigningCredentialNone {
    /**
     * 
     * @type {Web3SigningCredentialType}
     * @memberof Web3SigningCredentialNone
     */
    type: Web3SigningCredentialType;
}
/**
 * 
 * @export
 * @interface Web3SigningCredentialPrivateKeyHex
 */
export interface Web3SigningCredentialPrivateKeyHex {
    /**
     * 
     * @type {Web3SigningCredentialType}
     * @memberof Web3SigningCredentialPrivateKeyHex
     */
    type: Web3SigningCredentialType;
    /**
     * The ethereum account (public key) that the credential belongs to. Basically the username in the traditional terminology of authentication.
     * @type {string}
     * @memberof Web3SigningCredentialPrivateKeyHex
     */
    ethAccount: string;
    /**
     * The HEX encoded private key of an eth account.
     * @type {string}
     * @memberof Web3SigningCredentialPrivateKeyHex
     */
    secret: string;
}
// ...
export interface WithdrawRequest {
    /**
     * Contract locked id
     * @type {string}
     * @memberof WithdrawRequest
     */
    id: string;
    /**
     * Secret need to unlock the contract
     * @type {string}
     * @memberof WithdrawRequest
     */
    secret: string;
    /**
     * Contract address
     * @type {string}
     * @memberof WithdrawRequest
     */
    contractAddress: string;
    /**
     * 
     * @type {Web3SigningCredential}
     * @memberof WithdrawRequest
     */
    web3SigningCredential: Web3SigningCredential;
    /**
     * connectorId for the connector besu plugin
     * @type {string}
     * @memberof WithdrawRequest
     */
    connectorId: string;
    /**
     * 
     * @type {string | number}
     * @memberof WithdrawRequest
     */
    gas?: string | number;
}